widgets: rename mixins to types
