+++ lektor/admin/static/js/widgets/types.ts
